修改

+++ app/src/main/java/com/mt/androidtest_as/MyListViewFragment.java
@@ -3,12 +3,14 @@ package com.mt.androidtest_as;
 import android.app.Activity;
 import android.app.Dialog;
 import android.content.DialogInterface;
+import android.graphics.drawable.Drawable;
 import android.os.Bundle;
 import android.support.v7.app.AlertDialog;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ListView;
+import android.widget.TextView;
 
 import com.mt.androidtest_as.data.BaseData;
 import com.mt.androidtest_as.data.DataBank;
@@ -23,6 +25,8 @@ import java.util.List;
 public class MyListViewFragment extends BaseFragment {
     private Activity mActivity = null;
     private ListView mListView = null;
+    private View emptyView = null;
+    private static final int emptyViewIndex = -1;
     private MyLvAdapter mAdapter = null;
     private List<BaseData> mData = null;
     private boolean ifSetAdapter = false;
@@ -37,7 +41,11 @@ public class MyListViewFragment extends BaseFragment {
         super.onCreateView(inflater, container, savedInstanceState);
         View v = inflater.inflate(R.layout.fragment_mylistview, container, false);
         mListView = (ListView)v.findViewById(R.id.my_listview);
+        emptyView = inflater.inflate(R.layout.item_empty_view, (ViewGroup) mListView.getParent(), false);
+        emptyView.setTag(MyLvAdapter.ConvertViewTagID,emptyViewIndex);
+        emptyView.setOnClickListener(this);
         mAdapter = new MyLvAdapter(this);
+        setListViewEmptyText(mListView);
         return v;
     }
 
@@ -60,6 +68,14 @@ public class MyListViewFragment extends BaseFragment {
         mAdapter.notifyDataSetChanged();
     }
 
+    /**
+     * 数据为空时提示用户增加数据
+     */
+    private void setListViewEmptyText(ListView mListView){
+        ((ViewGroup)mListView.getParent()).addView(emptyView);
+        mListView.setEmptyView(emptyView);
+    }
+
     @Override
     public void onCreateDataConfirm(int dialogDataIndex) {
         int count = (dialogDataIndex+1)*5;
@@ -78,6 +94,10 @@ public class MyListViewFragment extends BaseFragment {
     @Override
     public void onClick(View v) {
         position = (int)v.getTag(MyLvAdapter.ConvertViewTagID);
+        if(emptyViewIndex == position){
+            onCreateDataConfirm(0);
+            return;
+        }
         if(null == mDialog) {
             mDialog = new AlertDialog.Builder(mActivity)
                     .setTitle(mActivity.getString(R.string.delete_item))
